refactor: `sift_down` to use `std::span` for improved clarity.
Before, we were passing a vector and manually indicating where
to start and stop the sift down. With a std::span we can avoid
that entirely, as we only need to pass the appropriate window
of data, and the algorithm will do the rest. This means that
we can remove unnecessary checks.