fix(alg): problem with quick_sort if all elements are the same
In the previous implementation, we ended up in an infinite loop for cases
were all the values of the vector (or sub-vectors) were the same. To avoid this
we need to split the values into two instead of three sub-vectors: values
less than the random value, equal than, and greater than.